pyproject.toml
[tool.poetry]
name = "newjob"
version = "0.2.0"
description = "Backend do projeto NewJob, desenvolvido na Fábrica de Software em parceria com o programa de pós graduação produção e sanidade animal"
authors = ["Eduardo da Silva <eduardo.silva@ifc.edu.br>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.1.6"
djangorestframework = "^3.12.2"
drf-spectacular = "^0.13.1"
djangorestframework-simplejwt = "^4.6.0"
django-rest-registration = "^0.6.0"
python-magic = "^0.4.22"
Pillow = "^8.1.0"
rules = "^2.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.8.0"
pre-commit = "^2.12.0"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
bandit = "^1.7.0"
pytest-django = "^4.2.0"
model-bakery = "^1.2.1"

[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
